refactor: 스케쥴 도메인에 필요한 유저를 쿠키를 사용하여 정보 찾는 기능 구현

+++ src/main/java/com/schedule/schedule/service/ScheduleService.java
@@ -1,5 +1,6 @@
 package com.schedule.schedule.service;
 
+import com.schedule.schedule.dto.CreateScheduleRequestDto;
 import com.schedule.schedule.dto.ScheduleRequestDto;
 import com.schedule.schedule.dto.ScheduleResponseDto;
 import com.schedule.schedule.entity.Schedule;
@@ -27,8 +28,8 @@ public class ScheduleService {
     private final UserRepository userRepository;
 
     // 스케쥴 생성
-    public ScheduleResponseDto createSchedule(ScheduleRequestDto dto) {
-        User user = userRepository.findById(dto.getUserId()).orElseThrow();
+    public ScheduleResponseDto createSchedule(CreateScheduleRequestDto dto, long userId) {
+        User user = userRepository.findById(userId).orElseThrow();
         Schedule schedule = new Schedule(user, dto.getTitle(), dto.getContents());
         Schedule savedSchedule = scheduleRepository.save(schedule);
 
@@ -60,38 +61,53 @@ public class ScheduleService {
     }
 
     // 스케쥴 id로 조회
-    public ScheduleResponseDto findById(long id) {
+    public ScheduleResponseDto findById(long id, long userId) {
+        User user = userRepository.findById(id).orElseThrow();
         Schedule schedule = scheduleRepository.findById(id).orElseThrow();
-        return new ScheduleResponseDto(
-                schedule.getId(),
-                schedule.getTitle(),
-                schedule.getContents(),
-                schedule.getCreatedAt(),
-                schedule.getUpdatedAt(),
-                schedule.getUser().getId()
-        );
+        if(user.getId() == schedule.getUser().getId()) {
+            return new ScheduleResponseDto(
+                    schedule.getId(),
+                    schedule.getTitle(),
+                    schedule.getContents(),
+                    schedule.getCreatedAt(),
+                    schedule.getUpdatedAt(),
+                    schedule.getUser().getId()
+            );
+        } else {
+            throw new RuntimeException("조회 실패;");
+        }
     }
 
     // 스케쥴 업데이트
-    public ScheduleResponseDto updateById(long id, ScheduleRequestDto dto) {
+    public ScheduleResponseDto updateById(long id, ScheduleRequestDto dto, long userId) {
         Schedule schedule = scheduleRepository.findById(id).orElseThrow();
         schedule.update(dto.getTitle(), dto.getContents());
+        User user = userRepository.findById(id).orElseThrow();
+        if(user.getId() == schedule.getUser().getId()) {
+            scheduleRepository.save(schedule);
+            return new ScheduleResponseDto(
+                    schedule.getId(),
+                    schedule.getTitle(),
+                    schedule.getContents(),
+                    schedule.getCreatedAt(),
+                    schedule.getUpdatedAt(),
+                    schedule.getUser().getId()
+            );
+        } else {
+            throw new RuntimeException("업데이트 실패");
+        }
 
-        scheduleRepository.save(schedule);
-        return new ScheduleResponseDto(
-                schedule.getId(),
-                schedule.getTitle(),
-                schedule.getContents(),
-                schedule.getCreatedAt(),
-                schedule.getUpdatedAt(),
-                schedule.getUser().getId()
-        );
+
     }
 
-    public String deleteById(long id) {
+    public String deleteById(long id, long userId) {
+        User user = userRepository.findById(id).orElseThrow();
         Schedule schedule = scheduleRepository.findById(id).orElseThrow();
-        scheduleRepository.deleteById(id);
-
-        return schedule.getTitle();
+        if(user.getId() == schedule.getUser().getId()) {
+            scheduleRepository.deleteById(id);
+            return schedule.getTitle();
+        } else {
+            throw new RuntimeException("삭제 실패");
+        }
     }
 }
